+++ lib/src/telegram/telegram.dart
@@ -1059,7 +1059,7 @@ class Telegram {
   ///
   /// [unbanned]: https://core.telegram.org/bots/api#unbanchatmember
   Future<bool> kickChatMember(dynamic chat_id, int user_id,
-      {int until_date}) async {
+      {int until_date, bool revoke_messages}) async {
     if (chat_id is! String && chat_id is! int) {
       return Future.error(TelegramException(
           'Attribute \'chat_id\' can only be either type of String or int'));
@@ -1069,6 +1069,7 @@ class Telegram {
       'chat_id': chat_id,
       'user_id': user_id,
       'until_date': until_date,
+      'revoke_messages': revoke_messages,
     };
     return await HttpClient.httpPost(requestUrl, body: body);
   }
@@ -1128,14 +1129,16 @@ class Telegram {
   /// https://core.telegram.org/bots/api#promotechatmember
   Future<bool> promoteChatMember(dynamic chat_id, int user_id,
       {bool is_anonymous,
-      bool can_change_info,
+      bool can_manage_chat,
       bool can_post_messages,
       bool can_edit_messages,
       bool can_delete_messages,
-      bool can_invite_users,
+      bool can_manage_voice_chats,
       bool can_restrict_members,
-      bool can_pin_messages,
-      bool can_promote_members}) async {
+      bool can_promote_members,
+      bool can_change_info,
+      bool can_invite_users,
+      bool can_pin_messages}) async {
     if (chat_id is! String && chat_id is! int) {
       return Future.error(TelegramException(
           'Attribute \'chat_id\' can only be either type of String or int'));
@@ -1145,14 +1148,16 @@ class Telegram {
       'chat_id': chat_id,
       'user_id': user_id,
       'is_anonymous': is_anonymous,
-      'can_change_info': can_change_info,
+      'can_manage_chat': can_manage_chat,
       'can_post_messages': can_post_messages,
       'can_edit_messages': can_edit_messages,
       'can_delete_messages': can_delete_messages,
-      'can_invite_users': can_invite_users,
+      'can_manage_voice_chats': can_manage_voice_chats,
       'can_restrict_members': can_restrict_members,
-      'can_pin_messages': can_pin_messages,
       'can_promote_members': can_promote_members,
+      'can_change_info': can_change_info,
+      'can_invite_users': can_invite_users,
+      'can_pin_messages': can_pin_messages,
     };
     return await HttpClient.httpPost(requestUrl, body: body);
   }
@@ -1203,6 +1208,78 @@ class Telegram {
     var requestUrl = '${_baseUrl}${_token}/exportChatInviteLink';
     var body = <String, dynamic>{'chat_id': chat_id};
     return await HttpClient.httpPost(requestUrl, body: body);
+  }
+
+  /// Use this method to create an additional invite link for a chat.
+  /// The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+  /// The link can be revoked using the method [revokeChatInviteLink].
+  /// Returns the new invite link as [ChatInviteLink] object.
+  ///
+  /// https://core.telegram.org/bots/api#createchatinvitelink
+  ///
+  /// [revokeChatInviteLink]: https://core.telegram.org/bots/api#revokechatinvitelink
+  /// [ChatInviteLink]: https://core.telegram.org/bots/api#chatinvitelink
+  Future<ChatInviteLink> createChatInviteLink(dynamic chat_id,
+      {int expire_date, int member_limit}) async {
+    if (chat_id is! String && chat_id is! int) {
+      return Future.error(TelegramException(
+          'Attribute \'chat_id\' can only be either type of String or int'));
+    }
+    var requestUrl = '${_baseUrl}${_token}/createChatInviteLink';
+    var body = <String, dynamic>{
+      'chat_id': chat_id,
+      'expire_date': expire_date,
+      'member_limit': member_limit,
+    };
+    return ChatInviteLink.fromJson(
+        await HttpClient.httpPost(requestUrl, body: body));
+  }
+
+  /// Use this method to edit a non-primary invite link created by the bot.
+  /// The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+  /// Returns the edited invite link as a [ChatInviteLink] object.
+  ///
+  /// https://core.telegram.org/bots/api#editchatinvitelink
+  ///
+  /// [ChatInviteLink]: https://core.telegram.org/bots/api#chatinvitelink
+  Future<ChatInviteLink> editChatInviteLink(dynamic chat_id, String invite_link,
+      {int expire_date, int member_limit}) async {
+    if (chat_id is! String && chat_id is! int) {
+      return Future.error(TelegramException(
+          'Attribute \'chat_id\' can only be either type of String or int'));
+    }
+    var requestUrl = '${_baseUrl}${_token}/editChatInviteLink';
+    var body = <String, dynamic>{
+      'chat_id': chat_id,
+      'invite_link': invite_link,
+      'expire_date': expire_date,
+      'member_limit': member_limit,
+    };
+    return ChatInviteLink.fromJson(
+        await HttpClient.httpPost(requestUrl, body: body));
+  }
+
+  /// Use this method to revoke an invite link created by the bot. If the primary link is revoked,
+  /// a new link is automatically generated.
+  /// The bot must be an administrator in the chat for this to work and must have the appropriate admin rights.
+  /// Returns the revoked invite link as [ChatInviteLink] object.
+  ///
+  /// https://core.telegram.org/bots/api#revokechatinvitelink
+  ///
+  /// [ChatInviteLink]: https://core.telegram.org/bots/api#chatinvitelink
+  Future<ChatInviteLink> revokeChatInviteLink(
+      dynamic chat_id, String invite_link) async {
+    if (chat_id is! String && chat_id is! int) {
+      return Future.error(TelegramException(
+          'Attribute \'chat_id\' can only be either type of String or int'));
+    }
+    var requestUrl = '${_baseUrl}${_token}/revokeChatInviteLink';
+    var body = <String, dynamic>{
+      'chat_id': chat_id,
+      'invite_link': invite_link,
+    };
+    return ChatInviteLink.fromJson(
+        await HttpClient.httpPost(requestUrl, body: body));
   }
 
   /// Use this method to set a profile photo for the chat.